Make code more generic
