+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/bulk/print/helper/EmailTaskHelper.java
@@ -1,0 +1,32 @@
+package uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper;
+
+import com.google.common.collect.ImmutableMap;
+import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.TaskContext;
+
+import java.util.Map;
+
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_CCD_REFERENCE_KEY;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_PET_NAME;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_RESP_NAME;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_SOLICITOR_NAME;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getPetitionerFullName;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getPetitionerSolicitorFullName;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getRespondentFullName;
+import static uk.gov.hmcts.reform.divorce.orchestration.tasks.util.TaskUtils.getCaseId;
+
+public class EmailTaskHelper {
+    public static ImmutableMap<String, String> citizenTemplateVariables(Map<String, Object> caseData) {
+        return ImmutableMap.of(
+            NOTIFICATION_PET_NAME, getPetitionerFullName(caseData)
+        );
+    }
+
+    public static ImmutableMap<String, String> solicitorTemplateVariables(TaskContext taskContext, Map<String, Object> caseData) {
+        return ImmutableMap.of(
+            NOTIFICATION_PET_NAME, getPetitionerFullName(caseData),
+            NOTIFICATION_RESP_NAME, getRespondentFullName(caseData),
+            NOTIFICATION_CCD_REFERENCE_KEY, getCaseId(taskContext),
+            NOTIFICATION_SOLICITOR_NAME, getPetitionerSolicitorFullName(caseData)
+        );
+    }
+}
